Refactor

+++ backend/src/api/controller/BaseController.ts
@@ -1,54 +1,24 @@
 import { NextFunction, Request, Response } from "express";
+import { loadData } from "../../utils/model";
 import { SdkResponse } from "@descope/node-sdk";
 import bcrypt from "bcrypt";
 import connectDescope from "../../config/descope.config";
-import * as querystring from "querystring";
 import supabase from "../../config/db.config";
-import loadData from "../../utils/loadData";
+import * as querystring from "querystring";
 
 const profile = async (req: Request, res: Response) => {
     const [password, subject, ...rest] = req.user;
     res.status(200).json(rest);
 };
 
-const create = async (subject: string, table: string, obj: object) => {
-    const data = await loadData(subject, table);
-
-    if (data) throw new Error("User already exists");
-
-    const { error } = await supabase.from(table).insert([
-        {
-            subject: subject,
-            ...obj,
-        },
-    ]);
-
-    if (error) throw new Error("Internal Server Error");
-};
-
-const update = async (
-    subject: string,
-    attributes: object,
-    table: string,
-    validAttributes: string[]
-) => {
-    for (const key in attributes) {
-        let index = validAttributes.indexOf(key);
-        if (index === -1) throw new Error("Bad Request");
-    }
-
-    const { error } = await supabase
-        .from(table)
-        .update(attributes)
-        .eq("subject", subject);
-
-    if (error) throw new Error("Internal Server Error");
-};
-
-const del = async (req: Request, table: string) => {
+const del = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const email = querystring.escape(req.query.email as string);
+        const email = querystring.unescape(req.query.email as string);
+        const pattern =
+            /^[a-za-z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-za-z0-9](?:[a-za-z0-9-]{0,61}[a-za-z0-9])?(?:\.[a-za-z0-9](?:[a-za-z0-9-]{0,61}[a-za-z0-9])?)*$/;
+        const isValidEmail = pattern.test(email);
 
+        if (!email || !isValidEmail) throw new Error("Bad Request");
         if (!email) throw new Error("Bad Request");
 
         const descopeClient = connectDescope();
@@ -56,17 +26,19 @@ const del = async (req: Request, table: string) => {
         await descopeClient.management.user.delete(email);
 
         const { error } = await supabase
-            .from(table)
+            .from(req.user.role)
             .delete()
-            .eq("subject", req.subject);
+            .eq("id", req.user.id);
 
         if (error) throw new Error("Internal Server Error");
+
+        res.status(200).json({ message: "Deleted" });
     } catch (error) {
         throw new Error("Internal Server Error");
     }
 };
 
-const updateModelEmail = async (req: Request, table: string) => {
+const updateEmail = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const { email } = req.body;
         const pattern =
@@ -76,25 +48,65 @@ const updateModelEmail = async (req: Request, table: string) => {
         if (!email || !req.token || !isValidEmail)
             throw new Error("Bad Request");
 
-        const descopeClient = connectDescope();
-        await descopeClient.magicLink.update.email(
+        await connectDescope().magicLink.update.email(
             req.user.email,
             email,
             req.token
         );
 
         const { error } = await supabase
-            .from(table)
+            .from(req.user.role)
             .update({ email })
-            .eq("subject", req.subject);
+            .eq("id", req.user.id);
 
         if (error) throw new Error("Internal Server Error");
+
+        res.status(200).json({ message: "Email Successfully Updated" });
     } catch (error) {
-        throw new Error("Internal Server Error");
+        next(error);
     }
 };
 
-const setPassword = async (req: Request, table: string) => {
+const updatePassword = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+) => {
+    try {
+        await setPassword(req, (email: string, hash: string, token: string) => {
+            connectDescope()
+                .password.update(email, hash, token)
+                .catch(() => {
+                    throw new Error("Internal Server Error");
+                });
+        });
+        res.status(200).json({ message: "Password Updated Successfully" });
+    } catch (error) {
+        next(error);
+    }
+};
+
+const resetPassword = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+) => {
+    try {
+        await setPassword(req, (email: string, hash: string, token: string) => {
+            connectDescope()
+                .password.sendReset(email, "http://omni-auth.vercel.app/")
+                .then((response) => {})
+                .catch(() => {
+                    throw new Error("Internal Server Error");
+                });
+        });
+        res.status(200).json({ message: "Password Succesfully Changed" });
+    } catch (error) {
+        next(error);
+    }
+};
+
+const setPassword = async (req: Request, cb: Function) => {
     try {
         const { email } = req.user;
         const { password } = req.body;
@@ -114,11 +126,13 @@ const setPassword = async (req: Request, table: string) => {
             throw new Error("Internal Server Error");
         });
 
-        await connectDescope()
-            .password.update(email, hash, req.token)
-            .catch(() => {
-                throw new Error("Internal Server Error");
-            });
+        await cb(email, hash, req.token);
+        const { error } = await supabase
+            .from(req.user.role)
+            .update({ password: hash })
+            .eq("id", req.user.id);
+
+        if (error) throw new Error("Internal Server Error");
     } catch (error) {
         throw new Error("Internal Server Error");
     }
@@ -129,20 +143,25 @@ const validatePassword = async (
     res: Response,
     next: NextFunction
 ) => {
-    const { password } = req.user;
+    try {
+        const { password } = req.user;
 
-    if (Object.keys(req.body).length != 1) throw new Error("Bad Request");
-    if (!password) throw new Error("Not Found");
+        if (Object.keys(req.body).length != 1) throw new Error("Bad Request");
+        if (!password) throw new Error("Not Found");
 
-    await bcrypt
-        .compare(req.body.password, password)
-        .then((value) => {
-            if (!value) throw new Error("Unathorized");
-            return true;
-        })
-        .catch(() => {
-            throw new Error("Internal Server Error");
-        });
+        await bcrypt
+            .compare(req.body.password, password)
+            .then((value) => {
+                if (!value) throw new Error("Unathorized");
+                return true;
+            })
+            .catch(() => {
+                throw new Error("Internal Server Error");
+            });
+        res.status(200).json({ message: "Password Succesfully Validated" });
+    } catch (error) {
+        next(error);
+    }
 };
 
 const logout = async (req: Request, res: Response, next: NextFunction) => {
@@ -165,11 +184,10 @@ const logoutAll = async (req: Request, res: Response, next: NextFunction) => {
 
 export {
     profile,
-    create,
-    update,
     del,
-    updateModelEmail,
-    setPassword,
+    updateEmail,
+    updatePassword,
+    resetPassword,
     validatePassword,
     logout,
     logoutAll,
